pocketchip-menu: Add battery indicator (#2)
* pocketchip-menu: Add battery indicator

Add a battery indicator using the Python psutil library. This should
allow the code to be portable. Removed the old icons and updated with
a new set of icons based on the icon8 theme used for the applications.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* pocketchip-menu: Hide mouse cursor

This is to add unclutter to the startup script so that the mouse cursor
is hidden from view unless you are actively clicking or moving a mouse.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* pocketchip-menu: Add Wifi Status Icon

Add a status icon for wifi. For now this only shows the status of an
active network connection (which still needs to be made via nmcli
outside of the UI).

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* pocketchip-menu: Include new Wifi assets

Forgot to include these in the prior commit, they might be important.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* pocketchip-menu: Only redraw the UI every 5 seconds

This change reduces the frequency of UI refreshes from "as fast as
possible" to "once every 5 seconds" (to account for battery or wifi
indicator changes). This reduces the CPU overhead from around 70%
to a much more managable 7%.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>