main: Switch to fastevent
Change the program to use fastevent (instead of event) so we can
utilize multi-threading. Change the way we get events from wait
which pauses execution until an event is available to get which
returns a list of all available events (or an empty list) at every
cycle. Change the main event handler to deal with an event list
instead of an event object.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>