pocketchip-menu: main: overhaul page refresh process
Use the event queue to notify when the screen needs to be refreshed
and only then should we redraw the screen. This took my CPU utilization
down below 1 percent for the program at idle. Note I also changed the
refresh to 10FPS, but this can probably be adjusted as needed.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>