Overhaul Screen Refresh Process/Thread Widgets (#6)
* Settings: hook up backlight for settings page

This uses a simple os directory traversal in sysfs to find the
current value of the backlight, and it uses a dbus call to set the
value of the backlight (to avoid using root). I also changed the
slider class to avoid using floats.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* Add DBUS testing commands as I work out the wifi bits.

* main: Make main program executable

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* main: Switch to fastevent

Change the program to use fastevent (instead of event) so we can
utilize multi-threading. Change the way we get events from wait
which pauses execution until an event is available to get which
returns a list of all available events (or an empty list) at every
cycle. Change the main event handler to deal with an event list
instead of an event object.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* wifi: Make module not fail when cannot query networkmanager

On one of my test machines I have networkmanager, but do not have
it running. Make this condition non-fatal.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* widgets: Make widget updates backgrounded/threaded

Widgets will now update according to helper threads. Additionally they
send a signal to their parent that there was a change so the parent
can more intelligently refresh itself to improve performance.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* pocketchip-menu: main: overhaul page refresh process

Use the event queue to notify when the screen needs to be refreshed
and only then should we redraw the screen. This took my CPU utilization
down below 1 percent for the program at idle. Note I also changed the
refresh to 10FPS, but this can probably be adjusted as needed.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* main: Tweak screen refreshes for faster performance

Make the entire screen refresh process dependent on certain events.
This adds a bit more performance.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* test: Add additional helper functions to dbustest

Adds some more dbus functions to the dbustest.py file to test out
capturing battery information.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

* dbus: Switch All Widgets to DBus Listener Refresh

All widgets (Bluetooth, Battery and Wifi) now only update when DBus
signal handlers tell them an update was sent.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>